djangoapi/requirements.txt
django
djangorestframework
djangorestframework_simplejwt
psycopg2